requirements.txt
virtualenv==20.21.0
python-dotenv==1.0.0
selenium==4.8.3